feat: Split Core package into 2 packages Core and Core.AspNet
