+++ .github/workflows/pipeline.yaml
@@ -1,10 +1,24 @@
 name: main-ci
 
 on:
+  pull_request:
+    branches: [ "main" ]
   push:
     branches: [ "main" ]
 
   workflow_dispatch:
+    
+concurrency:
+  group: ${{ github.head_ref || github.run_id }}-build
+  cancel-in-progress: true
+
+permissions:
+  id-token: write
+  contents: write
+  checks: write
+  issues: read
+  pull-requests: write
+  actions: write
 
 jobs:
   build:
@@ -13,20 +27,7 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-      - name: Install GitVersion
-        uses: gittools/actions/gitversion/setup@v0.9.15
-        with:
-          versionSpec: '5.x'
-
-      - name: Checkout
-        uses: actions/checkout@v3
-        with:
-          fetch-depth: 0 # Needed for gitversion
-      
-      - name: Determine Version
-        uses: gittools/actions/gitversion/execute@v0.9.15
-        with:
-          useConfigFile: true
+      - uses: ShipitSmarter/actions/initialise@main
 
       - name: Setup .NET Core
         uses: actions/setup-dotnet@v3
@@ -41,12 +42,15 @@ jobs:
         run: dotnet test
 
       - name: Pack
-        run: dotnet pack src/ShipitSmarter.Core/ShipitSmarter.Core.csproj  --no-build --configuration Release -p:PackageVersion=${{ env.GitVersion_NuGetVersion }}
+        run: |
+          dotnet pack src/ShipitSmarter.Core/ShipitSmarter.Core.csproj  --no-build --configuration Release -p:PackageVersion=${{ env.GitVersion_NuGetVersion }}
+          dotnet pack src/ShipitSmarter.Core.AspNet/ShipitSmarter.Core.AspNet.csproj  --no-build --configuration Release -p:PackageVersion=${{ env.GitVersion_NuGetVersion }}
 
       - name: PushNuget
         run: dotnet nuget push **/*.nupkg --source https://api.nuget.org/v3/index.json --api-key ${{ secrets.SHIPITSMARTER_CORE_NUGET_API_KEY }} --skip-duplicate
       
       - name: Create Tag
+        if: github.ref_name == 'main'
         uses: negz/create-tag@v1
         with:
           version: ${{ env.GitVersion_NuGetVersion }}
@@ -54,8 +58,17 @@ jobs:
           token: ${{ secrets.GITHUB_TOKEN }}
           
       - name: Remove pull request versions of package
+        if: github.ref_name == 'main'
         uses: darenm/unlist-nuget@v1
         with:
           NUGET_PACKAGE: shipitsmarter.core # Full Package ID in lower case!
+          VERSION_REGEX: ^\d.\d.\d+-pullrequest.* # Regex pattern to match version
+          NUGET_KEY: ${{ secrets.SHIPITSMARTER_PACKAGE_DELETER }} # nuget.org API key
+
+      - name: Remove pull request versions of package
+        if: github.ref_name == 'main'
+        uses: darenm/unlist-nuget@v1
+        with:
+          NUGET_PACKAGE: shipitsmarter.core.aspnet # Full Package ID in lower case!
           VERSION_REGEX: ^\d.\d.\d+-pullrequest.* # Regex pattern to match version
           NUGET_KEY: ${{ secrets.SHIPITSMARTER_PACKAGE_DELETER }} # nuget.org API key
