+++ ShipitSmarter.Core.sln
@@ -4,6 +4,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ShipitSmarter.Core", "src\S
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ShipitSmarter.Core.Tests", "test\ShipitSmarter.Core.Tests\ShipitSmarter.Core.Tests.csproj", "{2E49D658-A526-406C-AE0E-A3440E6BDCE5}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ShipitSmarter.Core.AspNet", "src\ShipitSmarter.Core.AspNet\ShipitSmarter.Core.AspNet.csproj", "{4029F596-E24E-4174-8A0E-A120963FE086}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -18,5 +20,9 @@ Global
 		{2E49D658-A526-406C-AE0E-A3440E6BDCE5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2E49D658-A526-406C-AE0E-A3440E6BDCE5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2E49D658-A526-406C-AE0E-A3440E6BDCE5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4029F596-E24E-4174-8A0E-A120963FE086}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4029F596-E24E-4174-8A0E-A120963FE086}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4029F596-E24E-4174-8A0E-A120963FE086}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4029F596-E24E-4174-8A0E-A120963FE086}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
