+++ src/ShipitSmarter.Core.AspNet/Documentation/AbsoluteDateTimeOffsetSchemaFilter.cs
@@ -1,13 +1,12 @@
 using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
 
-namespace ShipitSmarter.Core.Documentation;
-
+namespace ShipitSmarter.Core.AspNet.Documentation;
 
 /// <summary>
 /// Swagger filter to add a clarification to DateTime schema to indicate that it is not allowed to specify a time zone
 /// </summary>
-public class LocalDateTimeSchemaFilter : ISchemaFilter
+public class AbsoluteDateTimeOffsetSchemaFilter : ISchemaFilter
 {
     /// <summary>
     /// Adds a clarification to DateTime schema to indicate that it is not allowed to specify a time zone
@@ -16,11 +15,11 @@ public class LocalDateTimeSchemaFilter : ISchemaFilter
     /// <param name="context"></param>
     public void Apply(OpenApiSchema schema, SchemaFilterContext context)
     {
-        if (context.Type != typeof(DateTime) && context.Type != typeof(DateTime?))
+        if (context.Type != typeof(DateTimeOffset) && context.Type != typeof(DateTimeOffset?))
         {
             return;
         }
 
-        schema.Description += ". NOTE: Time zone specification is not allowed";
+        schema.Description += ". NOTE: Time zone specification is mandatory";
     }
 }
