+++ src/ShipitSmarter.Core.AspNet/Documentation/LocalDateTimeSchemaFilter.cs
@@ -1,12 +1,13 @@
 using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
 
-namespace ShipitSmarter.Core.Documentation;
+namespace ShipitSmarter.Core.AspNet.Documentation;
+
 
 /// <summary>
 /// Swagger filter to add a clarification to DateTime schema to indicate that it is not allowed to specify a time zone
 /// </summary>
-public class AbsoluteDateTimeOffsetSchemaFilter : ISchemaFilter
+public class LocalDateTimeSchemaFilter : ISchemaFilter
 {
     /// <summary>
     /// Adds a clarification to DateTime schema to indicate that it is not allowed to specify a time zone
@@ -15,11 +16,11 @@ public class AbsoluteDateTimeOffsetSchemaFilter : ISchemaFilter
     /// <param name="context"></param>
     public void Apply(OpenApiSchema schema, SchemaFilterContext context)
     {
-        if (context.Type != typeof(DateTimeOffset) && context.Type != typeof(DateTimeOffset?))
+        if (context.Type != typeof(DateTime) && context.Type != typeof(DateTime?))
         {
             return;
         }
 
-        schema.Description += ". NOTE: Time zone specification is mandatory";
+        schema.Description += ". NOTE: Time zone specification is not allowed";
     }
 }
