+++ src/ShipitSmarter.Core.AspNet/ShipitSmarter.Core.AspNet.csproj
@@ -1,13 +1,13 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
     <PropertyGroup>
         <TargetFramework>net8.0</TargetFramework>
         <LangVersion>latest</LangVersion>
-        <Nullable>enable</Nullable>
-        <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
         <ImplicitUsings>enable</ImplicitUsings>
+        <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+        <Nullable>enable</Nullable>
         <Version>0.0.1-alpha.1</Version>
-        <PackageId>ShipitSmarter.Core</PackageId>
+        <PackageId>ShipitSmarter.Core.AspNet</PackageId>
         <Title>ShipitSmarter.Core</Title>
         <Authors>ShipItSmarter</Authors>
         <PackageReadmeFile>README.md</PackageReadmeFile>
@@ -15,16 +15,24 @@
         <PackageLicenseExpression>MIT</PackageLicenseExpression>
         <PackageProjectUrl>https://github.com/ShipitSmarter/ShipItSmarter.Core</PackageProjectUrl>
         <Description>A set of core functionalities used across ShipItSmarter SDK and functionalities</Description>
-        <PackageTags>ShipitSmarter, Core</PackageTags>
+        <PackageTags>ShipitSmarter, Core, AspNet</PackageTags>
     </PropertyGroup>
-    
+
+    <ItemGroup>
+        <FrameworkReference Include="Microsoft.AspNetCore.App" />
+    </ItemGroup>
+
     <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
-      <DocumentationFile>bin\Debug\ShipitSmarter.Core.xml</DocumentationFile>
+        <DocumentationFile>bin\Debug\ShipitSmarter.Core.AspNet.xml</DocumentationFile>
     </PropertyGroup>
 
     <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
-      <DocumentationFile>bin\Release\ShipitSmarter.Core.xml</DocumentationFile>
+        <DocumentationFile>bin\Release\ShipitSmarter.Core.AspNet.xml</DocumentationFile>
     </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Swashbuckle.AspNetCore.SwaggerGen" Version="6.5.0" />
+    </ItemGroup>
 
     <ItemGroup>
         <None Include="../../README.md" Pack="true" PackagePath="/">
@@ -33,11 +41,8 @@
     </ItemGroup>
 
     <ItemGroup>
-      <PackageReference Include="Ardalis.SmartEnum" Version="2.1.0" />
-      <PackageReference Include="JsonSchema.Net" Version="5.4.3" />
-      <PackageReference Include="Microsoft.Extensions.FileSystemGlobbing" Version="8.0.0" />
-      <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="8.0.0" />
-      <PackageReference Include="System.Text.Json" Version="8.0.0" />
-      <PackageReference Include="YamlDotNet" Version="13.7.1" />
+      <ProjectReference Include="..\ShipitSmarter.Core\ShipitSmarter.Core.csproj" />
     </ItemGroup>
+    
+
 </Project>
